댓글 수정 ajax 구현

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -324,7 +324,7 @@ function doModifyReply(replyId){
 			</div>
 
 			<div class="chat-bubble" id="reply-body-${reply.id }">${reply.body}</div>
-			<form class="chat-bubble" action="/usr/reply/doModify" style="display: none;" id="modify-form-${reply.id }">
+			<form class="chat-bubble" action="#" style="display: none;" id="modify-form-${reply.id }" onsubmit="return false">
 				<input name="reply-text-${reply.id }" type="text" value="${reply.body }" />
 			</form>
 
@@ -335,11 +335,9 @@ function doModifyReply(replyId){
 					<button onclick="doModifyReply('${reply.id }');" id="save-btn-${reply.id }" style="display: none;">저장</button>
 				</c:if>
 				<c:if test="${reply.userCanDelete }">
-					<a href="../reply/doDelete?id=${reply.id}&articleId=${article.id}">삭제</a>
+					<a onclick="if(confirm('정말 삭제?') == false) return false;" href="../reply/doDelete?id=${reply.id}&articleId=${article.id}">삭제</a>
 				</c:if>
 			</div>
-
-
 		</div>
 	</c:forEach>
 	<c:if test="${empty replies }">
